+++ NEWS.md
@@ -22,4 +22,5 @@
 ### Bugfixes
 
 * Fixed errors when uploading files/directory names with invalid characters (Pro #698)
+* Added error when rsession may be running a different version of R than expected (Pro #2477)
 * Fixed "No such file or directory" errors when auto-saving R Notebook chunks while running them (#9284)